Backend/requirements.txt
Flask
flask-cors
python-dotenv
pymongo
pytest